Fixed display and logic of post and get
Fixed forms and logic of post and get, as well as other templates.

+++ engine/post.php
@@ -7,19 +7,18 @@ if (DEBUG_VIS == 1) {
 $c = connect();
 $c->set_charset('utf8');
 
-$usuario = $c->real_escape_string($_POST['usuario']);
+// Checkeamos TODAS vars. Usamos las seguras para meter a BdD, pero las inseguras para pantalla, para evitar barras (\).
+if (isset($_POST['usuario'])) $usuario = $c->real_escape_string($_POST['usuario']);
+if (isset($_POST['content'])) $text    = $c->real_escape_string($_POST['content']);
+if (isset($_POST['title']))   $title   = $c->real_escape_string($_POST['title']);
+if (isset($_POST['state']))   $state   = $c->real_escape_string($_POST['state']);
+if (isset($_POST['post_id'])) $post_id = $c->real_escape_string($_POST['post_id']);
+if (isset($_POST['id']))      $id      = $c->real_escape_string($_POST['id']);
+if (isset($_POST['value']))   $value   = $c->real_escape_string($_POST['value']);
 
-
-//Checkeamos vars. Usamos las seguras para meter a BdD, pero las inseguras para pantalla, para evitar barras (\).
-$text  = $c->real_escape_string($_POST['text']);
-$title = $c->real_escape_string($_POST['title']);
-$state = $c->real_escape_string($_POST['state']);
-$post_id = ( $_POST['post_id'] ? $c->real_escape_string($_POST['post_id']) : FALSE);
-
-
-//cambiar el *texto* de una entrada
+/* 1. Cambiar el *contenido* de una entrada */
 if ($_POST['editorId']=="text") {
-	$query = "INSERT INTO posts (`id`, `post_id`, `title`, `author`, `text`, `date`, `prioridad`, `state`) VALUES (NULL, '";
+	$query = "INSERT INTO posts (`id`, `post_id`, `title`, `author`, `content`, `date`, `prioridad`, `state`) VALUES (NULL, '";
 	$query .= $post_id."', '";
 	$query .= $title."', '";
 	$query .= $_SESSION['nombre']."', '";
@@ -29,17 +28,15 @@ if ($_POST['editorId']=="text") {
 
 	$result = query($query,$c);
 	if (!$result) {
-	    die('Could not query:' . mysql_error());
+	    die('<p class="error">Error al intentar guardar el texto: ' . mysqli_error($c));
 	}elseif ($result){
-		echo (get_html($_POST['text']));
-	}else{
-		echo ("<p class\"error\">Lo siento, el texto no se ha guardado</p>");
+		echo (get_html($_POST['content']));
+    exit;
 	}
 
-
-//cambiar el *título* de una entrada
+/* 2. Cambiar el *título* de una entrada */
 }elseif ($_POST['editorId']=="posttitle"){
-	$query = "INSERT INTO posts (`id`, `post_id`, `title`, `author`, `text`, `date`, `prioridad`, `state`) VALUES (NULL, '";
+	$query = "INSERT INTO posts (`id`, `post_id`, `title`, `author`, `content`, `date`, `prioridad`, `state`) VALUES (NULL, '";
 	$query .= $post_id."', '";
 	$query .= $title."', '";
 	$query .= $_SESSION['nombre']."', '";
@@ -49,22 +46,23 @@ if ($_POST['editorId']=="text") {
 
 	$result = query($query,$c);
 	if (!$result) {
-	    die('<p class="error">Could not query:' . mysql_error()."<br />Intenta volver atrás.</p>");
+	    die('<p class="error">Error al intentar guardar el título: ' . mysqli_error($c));
 	}elseif ($result){
 		echo ($_POST['title']);
-	}else{
-		echo ("<p class=\"error\">Lo siento, el texto no se ha guardado</p>");
+    exit;
 	}
 
+/* 3. Formulario página nueva */
+}elseif (isset($_POST['check']) && $_POST['check']=="newpost"){
 
-/*formulario página nueva*/
-}elseif ($_POST['check']=="newpost"){
-	/*formulario página nueva*/
-	$query ="SELECT MAX( post_id ) FROM `posts`"; // tiene que haber una forma mejor de obtener la ID mas alta????
+  /* Averiguar la post_id mas alta para usar la siguente */
+	$query ="SELECT MAX( post_id ) FROM `posts`";
 	$result = query($query,$c);
-	$out = fetch_array($result);
-	$post_id = $out[0] + 1;
-	$query = "INSERT INTO posts (`id`, `post_id`, `title`, `author`, `text`, `date`, `prioridad`, `state`) VALUES (NULL, '";
+  $out = fetch_array($result);
+	$post_id = current($out) + 1;
+
+  /* insertar los nuevos datos */
+	$query = "INSERT INTO posts (`id`, `post_id`, `title`, `author`, `content`, `date`, `prioridad`, `state`) VALUES (NULL, '";
 	$query .= $post_id."', '";
 	$query .= $title."', '";
 	$query .= $_SESSION['nombre']."', '";
@@ -74,65 +72,64 @@ if ($_POST['editorId']=="text") {
 
 	$result = query($query,$c);
 	if (!$result) {
-	    die('Could not query:' . mysql_error());
+	    die('<p class="error">Error al intentar guardar la página: ' . mysqli_error($c));
 	}elseif ($result){ //Exito
     header("Location: " . ORIGIN . ROOT . $_POST[title]);
 		exit;
-	}else{
-		echo ("<p class\"error\">Lo siento, el texto no se ha guardado</p>");
 	}
 
-/*Borrado de una entrada. Cómo ser más seguro? */
-}elseif ($_POST['function']=="borrar"){
+/* 4. Borrado de una entrada. Cómo ser más seguro? */
+}elseif (isset($_POST['function']) && $_POST['function']=="borrar"){
 	$query ="DELETE FROM `posts` WHERE `post_id` =  '$_POST[post_id]'";
 	$result = query($query,$c);
 	if (!$result) {
-	    die('Could not query:' . mysql_error());
+	    die('<p class="error">Error al intentar borrar las entradas: ' . mysqli_error($c));
 	}elseif ($result){ //Exito
 		$_SESSION['title_del']=$_POST['title'];
 		header("Location: " . ORIGIN . ROOT);
 		exit;
 	}
 
-/*Cambio de estado */
-}elseif ($_POST['editorId']=="estado"){//Verificar valor de $_POST[value] y $_POST[id] antes de introducirlos?
-	$status = array ('P'=>"Planteada",'E'=>"En curso",'X'=>"Estancada",'F'=>"Esperando feedback",'C'=>"Cancelada",'H'=>"Hibernando");
-	$stat_flag = $_POST['value'];
+/* 5. Cambio de *estado* */
+}elseif ($_POST['editorId']=="estado"){
+	$status = array ('P'=>"Planteada",
+                   'E'=>"En curso",
+                   'X'=>"Estancada",
+                   'F'=>"Esperando feedback",
+                   'C'=>"Cancelada",
+                   'H'=>"Hibernando");
+	$stat_flag = $value;
 	if ($stat_flag!=""){ $state="Marcada como <span class=\"$stat_flag\">$status[$stat_flag]</span> | ";}else{$state="Marcar estado | ";}
-	$query="UPDATE `posts` SET `author` = '$_SESSION[nombre]', `state` = '$_POST[value]', `date`= NOW() WHERE `id` = $_POST[id]";
+	$query="UPDATE `posts` SET `author` = '$_SESSION[nombre]', `state` = '$value', `date`= NOW() WHERE `id` = $id";
 
 	$result = query($query,$c);
 	if (!$result) {
-	    die('<p class="error">Could not query:' . mysql_error()."<br />Intenta volver atrás.</p>");
+	    die('<p class="error">Error al intentar cambiar el estado: ' . mysqli_error($c)."<br />Intenta volver atrás.</p>");
 	}elseif ($result){ //Exito
 		echo $state;
-	}else{
-		echo ("<p class=\"error\">Lo siento, el texto no se ha guardado</p>");
+    exit;
 	}
 
-
-
-/* Cambio de relevancia */
-}elseif ($_POST['editorId']=="prioridad"){//Verificar valor de $_POST[value] y $_POST[id] antes de introducirlos?
-	if ($_POST['value']==1) {$impor="<span style=\"color:#f00;\">✔</span> <b>Importante</b> | ";}elseif($_POST['value']==0){$impor="Marcar prioridad | ";}
-	$query="UPDATE `posts` SET `author` = '$_SESSION[nombre]', `prioridad` =  '$_POST[value]', `date` = NOW() WHERE `id` = $_POST[id]";
+/* 6. Cambio de *prioridad* */
+}elseif ($_POST['editorId']=="prioridad"){
+	if ($value==1) {$impor="<span style=\"color:#f00;\">✔</span> <b>Importante</b> | ";}elseif($value==0){$impor="Marcar prioridad | ";}
+	$query="UPDATE `posts` SET `author` = '$_SESSION[nombre]', `prioridad` =  '$value', `date` = NOW() WHERE `id` = $id";
 
 	$result = query($query,$c);
 	if (!$result) {
 	    /*die*/ echo('<p class="error">Could not query:' . mysql_error()."<br />Intenta volver atrás.</p>");
 	}elseif ($result){ //Exito
 		echo $impor;
-	}else{
-		echo ("<p class=\"error\">Lo siento, el texto no se ha guardado</p>");
+    exit;
 	}
 
 }
 
 close($c);
 
-//debug
+/* debug */
 if (DEBUG_VIS == 1 && DEBUG_LVL == 2) {
-  debug_add("<pre>\n $query " . print_r ($_POST) . "\n</pre>\n");
+  debug_add("<pre>\n $query " . print_r ($_POST, TRUE) . "\n</pre>\n");
 }
 /**/
